modules/messages/pyproject.toml
[tool.poetry]
name = "messages"
version = "0.1.0"
description = ""
authors = ["Stefano Loscalzo <stefano.loscalzo@gmail.com>", "Davide Aimar <davide.aimar99gmail.com>", "Gabriele Belluardo <gabriele.belluardo@outlook.it>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
bob-core = { git = "https://github.com/policumbent/bob-core", rev = "main" }

# package deps
python = "^3.7.3"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
autopep8 = "^1.6.0"
pytest = "^6.2.5"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
